requirements.txt
fastapi
uvicorn
openai
npm
sqlalchemy
passlib
bcrypt==3.2.0
python-dotenv
python-jose
passlib
psycopg2